+++ projects/burnr/src/hooks/api/useApiCreate.ts
@@ -1,11 +1,13 @@
 import { useEffect, useState } from "react"
 import { ApiPromise } from "@polkadot/api"
 import { logger } from "@polkadot/util"
-import { createPolkadotJsScClient, WellKnownChain } from "@substrate/connect"
+import {
+  ScProvider,
+  WellKnownChain,
+} from "@polkadot/rpc-provider/substrate-connect"
 import { NETWORK, BURNR_WALLET } from "../../utils/constants"
 import { useIsMountedRef } from "./useIsMountedRef"
 
-const scClient = createPolkadotJsScClient()
 const l = logger(BURNR_WALLET)
 
 export const useApiCreate = (): ApiPromise => {
@@ -17,9 +19,8 @@ export const useApiCreate = (): ApiPromise => {
   useEffect((): void => {
     const choseSmoldot = async (endpoint: string): Promise<void> => {
       try {
-        const provider = await scClient.addWellKnownChain(
-          endpoint as WellKnownChain,
-        )
+        const provider = new ScProvider(endpoint as WellKnownChain)
+        await provider.connect()
         const api = await ApiPromise.create({ provider })
         l.log(`Burnr is now connected to ${endpoint}`)
         mountedRef.current && setApi(api)